requirements.txt
certifi==2017.4.17
chardet==3.0.4
idna==2.5
py==1.4.34
pytest==3.1.2
python-redmine==2.0.2
PyYAML==3.12
requests==2.18.1
urllib3==1.21.1
